rename main_ to core
